♻️refactor : 불필요하게 분리된 카드 컴포넌트 통합
