--- src/ui/SharedPossibleCard/index.js
@@ -1,1 +1,0 @@
-export * from "./SharedPossibleCard";
