+++ src/ui/WishListButton/WishListButton.css
@@ -1,11 +1,6 @@
-.WishListButton {
-  width: 2.125rem;
-  height: 2.125rem;
-}
-
 .WishListButton-button {
-  width: 2.125rem;
-  height: 2.125rem;
+  width: 2.5rem;
+  height: 2.5rem;
   position: absolute;
   top: 1.5rem;
   right: 1.5rem;
